// src/webparts/skillsToPptCv/components/SkillsToPptCv.module.scss.d.ts
declare const styles: {
    skillsToPptCv: string;
    container: string;
    row: string;
    column: string;
    'ms-Grid': string;
    title: string;
    subTitle: string;
    description: string;
    textBox: string;
    button: string;
    label: string;
    list: string;
    listItem: string;
};
export default styles;